+++ packages/stream_video/lib/src/call/call.dart
@@ -46,6 +46,7 @@ import '../utils/subscriptions.dart';
 import '../webrtc/media/media_constraints.dart';
 import '../webrtc/model/rtc_video_dimension.dart';
 import '../webrtc/model/rtc_video_parameters.dart';
+import '../webrtc/rtc_audio_api/rtc_audio_api.dart' as rtc_audio;
 import '../webrtc/rtc_manager.dart';
 import '../webrtc/rtc_media_device/rtc_media_device.dart';
 import '../webrtc/rtc_media_device/rtc_media_device_notifier.dart';
@@ -87,6 +88,7 @@ const _idConnect = 6;
 const _idAwait = 7;
 const _idFastReconnectTimeout = 8;
 const _idReconnect = 9;
+const _idNativeWebRtc = 10;
 
 const _tag = 'SV:Call';
 int _callSeq = 1;
@@ -381,10 +383,18 @@ class Call {
       _observeState();
       _observeReconnectEvents();
       _observeUserId();
+      _observeNativeWebRtcEventStream();
 
       _logger.v(() => '[_init] initialized');
       _initialized = true;
     });
+  }
+
+  void _observeNativeWebRtcEventStream() {
+    _subscriptions.add(
+      _idNativeWebRtc,
+      _onNativeWebRtcEvent(),
+    );
   }
 
   void _observeState() {
@@ -456,6 +466,52 @@ class Call {
         state.settings.audio.opusDtxEnabled;
     _sessionFactory.sdpEditor.opusRedEnabled =
         state.settings.audio.redundantCodingEnabled;
+  }
+
+  StreamSubscription<NativeWebRtcEvent> _onNativeWebRtcEvent() {
+    return RtcMediaDeviceNotifier.instance.nativeWebRtcEventsStream().listen((
+      event,
+    ) {
+      _logger.d(
+        () => '[_onNativeWebRtcEvent] screenSharingStopped: $event',
+      );
+
+      switch (event) {
+        case ScreenSharingStoppedEvent _:
+          if (CurrentPlatform.isIos) {
+            // On iOS only one broadcast extension can be active at a time
+            setScreenShareEnabled(enabled: false);
+          } else {
+            final trackId = event.data?['trackId'] as String?;
+            if (trackId != null && state.value.localParticipant != null) {
+              final track = getTrack(
+                state.value.localParticipant!.trackIdPrefix,
+                SfuTrackType.screenShare,
+              );
+
+              if (track?.mediaTrack.id == trackId) {
+                setScreenShareEnabled(enabled: false);
+              }
+            }
+          }
+          break;
+        case ScreenSharingStartedEvent _:
+          _stateManager.participantSetScreenShareEnabled(
+            enabled: true,
+          );
+
+          _connectOptions = _connectOptions.copyWith(
+            screenShare: TrackOption.enabled(
+              constraints: const ScreenShareConstraints(
+                useiOSBroadcastExtension: true,
+              ),
+            ),
+          );
+          break;
+        default:
+          return;
+      }
+    });
   }
 
   Future<void> _onCoordinatorEvent(StreamCallEvent event) async {
@@ -556,10 +612,15 @@ class Call {
       case StreamCallMissedEvent _:
         return _stateManager.callMetadataChanged(event.metadata);
       case StreamCallSessionEndedEvent _:
-        return _stateManager.callMetadataChanged(event.metadata);
+        return _stateManager.callMetadataChanged(
+          event.metadata,
+          updateMembers: false,
+        );
       case StreamCallSessionStartedEvent _:
-        return _stateManager.callMetadataChanged(event.metadata);
-
+        return _stateManager.callMetadataChanged(
+          event.metadata,
+          updateMembers: false,
+        );
       default:
         break;
     }
@@ -1906,15 +1967,21 @@ class Call {
     if (cameraOption is TrackProvided) {
       return _setLocalTrack(cameraOption.track);
     } else if (cameraOption is TrackEnabled) {
+      final constraints = cameraOption.constraints is CameraConstraints
+          ? cameraOption.constraints as CameraConstraints?
+          : null;
+
       return setCameraEnabled(
         enabled: true,
-        constraints: CameraConstraints(
-          facingMode: facingMode,
-          deviceId: deviceId,
-          params:
-              targetResolution?.toVideoParams() ??
-              RtcVideoParametersPresets.h720_16x9,
-        ),
+        constraints:
+            constraints ??
+            CameraConstraints(
+              facingMode: facingMode,
+              deviceId: deviceId,
+              params:
+                  targetResolution?.toVideoParams() ??
+                  RtcVideoParametersPresets.h720_16x9,
+            ),
       );
     }
 
@@ -1925,7 +1992,10 @@ class Call {
     if (microphoneOption is TrackProvided) {
       await _setLocalTrack(microphoneOption.track);
     } else if (microphoneOption is TrackEnabled) {
-      await setMicrophoneEnabled(enabled: true);
+      final constraints = microphoneOption.constraints is AudioConstraints
+          ? microphoneOption.constraints as AudioConstraints?
+          : null;
+      await setMicrophoneEnabled(enabled: true, constraints: constraints);
     }
   }
 
@@ -1936,15 +2006,22 @@ class Call {
     if (screenShareOption is TrackProvided) {
       await _setLocalTrack(screenShareOption.track);
     } else if (screenShareOption is TrackEnabled) {
+      final constraints =
+          screenShareOption.constraints is ScreenShareConstraints
+          ? screenShareOption.constraints as ScreenShareConstraints?
+          : null;
+
       await setScreenShareEnabled(
         enabled: true,
-        constraints: ScreenShareConstraints(
-          params:
-              targetResolution?.toVideoParams(
-                defaultBitrate: RtcVideoParametersPresets.k1080pBitrate,
-              ) ??
-              RtcVideoParametersPresets.h1080_16x9,
-        ),
+        constraints:
+            constraints ??
+            ScreenShareConstraints(
+              params:
+                  targetResolution?.toVideoParams(
+                    defaultBitrate: RtcVideoParametersPresets.k1080pBitrate,
+                  ) ??
+                  RtcVideoParametersPresets.h1080_16x9,
+            ),
       );
     }
   }
@@ -1962,13 +2039,19 @@ class Call {
       final mediaConstraints = track.mediaConstraints;
       if (mediaConstraints is AudioConstraints) {
         _logger.v(() => '[setLocalTrack]: setMicrophoneEnabled true');
-        await setMicrophoneEnabled(enabled: true);
+        await setMicrophoneEnabled(
+          enabled: true,
+          constraints: mediaConstraints,
+        );
       } else if (mediaConstraints is CameraConstraints) {
         _logger.v(() => '[setLocalTrack]: setCameraEnabled true');
-        await setCameraEnabled(enabled: true);
+        await setCameraEnabled(enabled: true, constraints: mediaConstraints);
       } else if (mediaConstraints is ScreenShareConstraints) {
         _logger.v(() => '[setLocalTrack] setScreenShareEnabled true');
-        await setScreenShareEnabled(enabled: true);
+        await setScreenShareEnabled(
+          enabled: true,
+          constraints: mediaConstraints,
+        );
       } else {
         streamLog.e(
           _tag,
@@ -2607,11 +2690,14 @@ class Call {
             .firstOrNull;
 
         _connectOptions = connectOptions.copyWith(
-          cameraFacingMode: connectOptions.cameraFacingMode.flip(),
+          cameraFacingMode: success.data.mediaConstraints.facingMode,
           videoInputDevice: currentInput,
         );
 
-        _stateManager.participantFlipCamera(currentInput);
+        _stateManager.participantFlipCamera(
+          currentInput,
+          track: success.data,
+        );
       },
       failure: (failure) {},
     );
@@ -2706,11 +2792,20 @@ class Call {
         Result.error('Session is null');
 
     if (result.isSuccess) {
-      _connectOptions = connectOptions.copyWith(videoInputDevice: device);
-      _stateManager.participantSetVideoInputDevice(device: device);
+      final track = result.getDataOrNull()!;
+
+      _connectOptions = connectOptions.copyWith(
+        videoInputDevice: device,
+        cameraFacingMode: track.mediaConstraints.facingMode,
+      );
+
+      _stateManager.participantSetVideoInputDevice(
+        device: device,
+        track: track,
+      );
     }
 
-    return result;
+    return result.map((_) => none);
   }
 
   Future<Result<None>> setCameraEnabled({
@@ -2748,9 +2843,17 @@ class Call {
         iOSMultitaskingCameraAccessEnabled: multitaskingEnabled,
       );
 
+      var facingMode = constraints?.facingMode;
+      if (facingMode == null && result.getDataOrNull() is RtcLocalCameraTrack) {
+        final track = result.getDataOrNull()! as RtcLocalCameraTrack;
+        facingMode = track.mediaConstraints.facingMode;
+      }
+
       _connectOptions = _connectOptions.copyWith(
-        camera: enabled ? TrackOption.enabled() : TrackOption.disabled(),
-        cameraFacingMode: constraints?.facingMode ?? FacingMode.user,
+        camera: enabled
+            ? TrackOption.enabled(constraints: constraints)
+            : TrackOption.disabled(),
+        cameraFacingMode: facingMode ?? _connectOptions.cameraFacingMode,
       );
     }
 
@@ -2811,7 +2914,9 @@ class Call {
       );
 
       _connectOptions = _connectOptions.copyWith(
-        microphone: enabled ? TrackOption.enabled() : TrackOption.disabled(),
+        microphone: enabled
+            ? TrackOption.enabled(constraints: constraints)
+            : TrackOption.disabled(),
       );
     }
 
@@ -2848,16 +2953,27 @@ class Call {
         ) ??
         Result.error('Call session is null, cannot start screen share');
 
+    // In case of iOS Broadcast Extension, we don't update the state here
+    // We listen to the ScreenShareStarted event instead
+    if (CurrentPlatform.isIos &&
+        constraints is ScreenShareConstraints &&
+        constraints.useiOSBroadcastExtension) {
+      return result.map((_) => none);
+    }
+
     if (result.isSuccess) {
       _stateManager.participantSetScreenShareEnabled(
         enabled: enabled,
       );
 
       _connectOptions = _connectOptions.copyWith(
-        screenShare: enabled ? TrackOption.enabled() : TrackOption.disabled(),
+        screenShare: enabled
+            ? TrackOption.enabled(constraints: updatedConstraints)
+            : TrackOption.disabled(),
       );
 
       if (enabled) {
+        // [web only] Automatically stop screen share when the track ends
         result.getDataOrNull()?.mediaTrack.onEnded = () {
           setScreenShareEnabled(enabled: false);
         };
@@ -2891,6 +3007,12 @@ class Call {
     }
   }
 
+  /// Sets the audio output device for the call.
+  /// - [device]: The audio output device to set.
+  /// Returns a [Result] indicating the success or failure of the operation.
+  ///
+  /// On web platforms, this method may return an error if the browser does not support
+  /// setting audio output devices programmatically.
   Future<Result<None>> setAudioOutputDevice(RtcMediaDevice device) async {
     final result =
         await _session?.setAudioOutputDevice(device) ??
@@ -2905,6 +3027,10 @@ class Call {
     }
 
     return result;
+  }
+
+  bool checkIfAudioOutputChangeSupported() {
+    return rtc_audio.checkIfAudioOutputChangeSupported();
   }
 
   /// Sets the mirror state for a remote participant's video track.