+++ packages/stream_video/lib/src/call/state/mixins/state_lifecycle_mixin.dart
@@ -111,18 +111,16 @@ mixin StateLifecycleMixin on StateNotifier<CallState> {
   }) {
     _logWithState('lifecycleCallCreated', 'ringing: $ringing');
 
-    state = state
-        .copyFromMetadata(
-          data.metadata,
-        )
-        .copyWith(
-          status: data.toCallStatus(state: state, ringing: ringing),
-          callParticipants: data.metadata.toCallParticipants(state),
-          isRingingFlow: ringing,
-          audioOutputDevice: callConnectOptions.audioOutputDevice,
-          audioInputDevice: callConnectOptions.audioInputDevice,
-          videoInputDevice: callConnectOptions.videoInputDevice,
-        );
+    final newState = state.copyFromMetadata(data.metadata);
+
+    state = newState.copyWith(
+      status: data.toCallStatus(state: newState, ringing: ringing),
+      callParticipants: data.metadata.toCallParticipants(newState),
+      isRingingFlow: ringing,
+      audioOutputDevice: callConnectOptions.audioOutputDevice,
+      audioInputDevice: callConnectOptions.audioInputDevice,
+      videoInputDevice: callConnectOptions.videoInputDevice,
+    );
   }
 
   void lifecycleCallRinging(
@@ -130,16 +128,14 @@ mixin StateLifecycleMixin on StateNotifier<CallState> {
   ) {
     _logWithState('lifecycleCallRinging');
 
-    state = state
-        .copyFromMetadata(
-          data.metadata,
-        )
-        .copyWith(
-          status: data.toCallStatus(state: state),
-          isRingingFlow: data.ringing,
-          ownCapabilities: data.metadata.details.ownCapabilities.toList(),
-          callParticipants: data.metadata.toCallParticipants(state),
-        );
+    final newState = state.copyFromMetadata(data.metadata);
+
+    state = newState.copyWith(
+      status: data.toCallStatus(state: newState),
+      isRingingFlow: data.ringing,
+      ownCapabilities: data.metadata.details.ownCapabilities.toList(),
+      callParticipants: data.metadata.toCallParticipants(newState),
+    );
   }
 
   void lifecycleCallJoining() {
@@ -157,18 +153,16 @@ mixin StateLifecycleMixin on StateNotifier<CallState> {
     final status = state.status.isJoining ? CallStatus.joined() : state.status;
     _logWithState('lifecycleCallJoined', 'newStatus: $status');
 
-    state = state
-        .copyFromMetadata(
-          data.metadata,
-        )
-        .copyWith(
-          status: status,
-          ownCapabilities: data.metadata.details.ownCapabilities.toList(),
-          callParticipants: data.metadata.toCallParticipants(state),
-          audioOutputDevice: callConnectOptions?.audioOutputDevice,
-          audioInputDevice: callConnectOptions?.audioInputDevice,
-          videoInputDevice: callConnectOptions?.videoInputDevice,
-        );
+    final newState = state.copyFromMetadata(data.metadata);
+
+    state = newState.copyWith(
+      status: status,
+      ownCapabilities: data.metadata.details.ownCapabilities.toList(),
+      callParticipants: data.metadata.toCallParticipants(newState),
+      audioOutputDevice: callConnectOptions?.audioOutputDevice,
+      audioInputDevice: callConnectOptions?.audioInputDevice,
+      videoInputDevice: callConnectOptions?.videoInputDevice,
+    );
   }
 
   void lifecycleCallReconnectingFailed() {