+++ packages/stream_video_push_notification/CHANGELOG.md
@@ -19,6 +19,12 @@ In this release, we removed the dependency on `flutter_callkit_incoming`, which 
 - `observeCallEndedCallKitEvent` → `observeCallEndedRingingEvent`
 - `CallKitEvent` (type) → `RingingEvent`
 
+## 0.11.2
+* Sync version with `stream_video_flutter` 0.11.2
+
+## 0.11.1
+* Sync version with `stream_video_flutter` 0.11.1
+
 ## 0.11.0
 
 🚧 Build breaking changes